Making the IGDB command only search for games
